+++ frontend/src/component/admin/traffic/Traffic.tsx
@@ -1,4 +1,4 @@
-import InstanceMetricsChart from '../instance-admin/InstanceMetrics/InstanceMetricsChart'; // TODO move this to traffic folder (pending conversation on how to organize this)
+import InstanceMetricsChart from './TrafficMetricsChart';
 import AdminMenu from '../menu/AdminMenu';
 
 export const Traffic = () => {
